refactor(drive-abci)!: drop the v14 validator-side rollback; port the 4.1.1 proposer fix
Remove process_raw_state_transitions v1 and DRIVE_ABCI_METHOD_VERSIONS_V10,
rewiring protocol v14 back to V9. Review of process_proposal established
that its unexpected_execution_results gate (present since early 2024)
already rejects any block whose execution yields InternalError or
UnpaidConsensusError results, so a block that includes a written-then-
failed transition can never commit — the v14 validator-side rollback
closed no reachable hole, only cleaned state inside a transaction the
reject path discards, while its genesis carve-out would have become a
permanent consensus rule (flagged as blocking in review).

In its place, forward-port the shipped 4.1.1 proposer-side rollback
(#4409) into v0, since v4.2-dev still carries the halt: savepoint per
executed transition while proposing at non-genesis heights, rollback on
drop-class results, exhaustive classification, and the fault-injection
hook (as the extracted maybe_override helper requested in review). The
halt regression tests now pin this path.

The breaking marker on this commit reflects reverting the earlier
v14 method-table change within this branch; net of the branch, the PR
no longer contains any consensus change.
