+++ packages/rs-drive-abci/src/execution/platform_events/state_transition_processing/process_raw_state_transitions/v0/mod.rs
@@ -17,9 +17,45 @@ use crate::platform_types::state_transitions_processing_result::{
 use dpp::util::hash::hash_single;
 use dpp::version::PlatformVersion;
 use drive::grovedb::Transaction;
+use drive::grovedb_storage::Error::RocksDBError;
 use std::time::Instant;
 
 use super::super::StateTransitionAwareError;
+
+/// Test-only fault injection: force the next successfully executed state transition to be
+/// reported as an `InternalError` AFTER its drive operations were applied. This models the
+/// only way an `InternalError` can carry state (an `Err` surfacing after
+/// `apply_drive_operations(apply = true)`, e.g. the address-input fee coverage guard failing
+/// on an under-estimated `Shield`) without depending on any particular estimation bug.
+#[cfg(test)]
+pub(crate) mod test_fault_injection {
+    use crate::platform_types::state_transitions_processing_result::StateTransitionExecutionResult;
+    use std::cell::Cell;
+
+    thread_local! {
+        pub static FAIL_NEXT_SUCCESSFUL_EXECUTION: Cell<bool> = const { Cell::new(false) };
+    }
+
+    /// If armed, consume the flag and replace a successful execution with the
+    /// `InternalError` a post-apply failure would produce; identity for every other result
+    /// and while unarmed. Kept here so the processing loop carries a single call instead of
+    /// the override logic.
+    pub(crate) fn maybe_override(
+        execution_result: StateTransitionExecutionResult,
+    ) -> StateTransitionExecutionResult {
+        if matches!(
+            execution_result,
+            StateTransitionExecutionResult::SuccessfulExecution { .. }
+        ) && FAIL_NEXT_SUCCESSFUL_EXECUTION.with(|flag| flag.replace(false))
+        {
+            StateTransitionExecutionResult::InternalError(
+                "injected post-apply failure (test_fault_injection)".to_string(),
+            )
+        } else {
+            execution_result
+        }
+    }
+}
 
 impl<C> Platform<C>
 where
@@ -69,6 +105,37 @@ where
         let state_transition_container =
             self.decode_raw_state_transitions(raw_state_transitions, platform_version)?;
 
+        // PROPOSER-SIDE ONLY (consensus-invisible, hence no protocol-version gate): while
+        // building a proposal, wrap each executed state transition in a savepoint and roll
+        // back if its result strips it from the block (`TxAction::Removed`). Execution can
+        // write into the shared block transaction before failing (the address-input fee flow
+        // is apply-then-check), and without the rollback the gossiped block omits the
+        // transition while the advertised app hash includes its writes — no validator can
+        // reproduce the hash, and every proposer carrying the transition burns its round
+        // (mainnet evo1 stalls of 2026-08-14/15, after heights 415652 and 415661).
+        //
+        // The validation path (`proposing_state_transitions == false`) is deliberately
+        // untouched: rolling back there would change what state a received block evaluates
+        // to, which is a consensus change needing a protocol-version gate — and one that is
+        // unnecessary, because `process_proposal` already REJECTS any block whose execution
+        // produced an `InternalError` or `UnpaidConsensusError` result (the
+        // `unexpected_execution_results` gate), so a block that carries such a transition
+        // can never commit and its writes die with the rejected round's transaction. This
+        // proposer-side rollback only changes which blocks this node BUILDS — the published
+        // block and app hash are exactly what any un-upgraded validator computes from that
+        // block, so mixed networks cannot diverge.
+        //
+        // The genesis height is excluded because its re-proposal path relies on a
+        // single-savepoint discipline: init_chain sets one savepoint, and each genesis round
+        // rewinds to it with one `rollback_to_savepoint()` (see prepare_proposal /
+        // process_proposal). Savepoints of KEPT transitions stay on the stack — RocksDB
+        // exposes no pop-without-rollback — and extra savepoints on the genesis transaction
+        // would redirect that rewind. At every other height each proposal round runs in a
+        // freshly started transaction that is either committed (leftover savepoints are inert
+        // markers) or dropped when the round ends, so the residue can affect nothing.
+        let rollback_dropped_transitions =
+            proposing_state_transitions && block_info.height != self.config.abci.genesis_height;
+
         let mut processing_result = StateTransitionsProcessingResult::default();
 
         for decoded_state_transition in state_transition_container.into_iter() {
@@ -108,6 +175,12 @@ where
                             );
                         }
 
+                        // Mark the state we can return to if this transition's result strips
+                        // it from the block (see `rollback_dropped_transitions` above).
+                        if rollback_dropped_transitions {
+                            transaction.set_savepoint();
+                        }
+
                         // Validate state transition and produce an execution event
                         let execution_result = process_state_transition(
                             &platform_ref,
@@ -117,9 +190,8 @@ where
                         )
                         .map(|validation_result| {
                             // Dispatch to the versioned helper (v0 = pre-v13, v1 = records
-                            // paid-invalid balance effects). Only this helper changed at v13, so this
-                            // outer loop stayed at v0 across that gate. (The loop itself changed at
-                            // v14 — see _v1, which rolls back dropped transitions.)
+                            // paid-invalid balance effects). Only this helper changed at v13; the outer
+                            // loop is version-agnostic, so it must NOT be versioned.
                             self.process_validation_result(
                                 raw_state_transition,
                                 &state_transition_name,
@@ -137,6 +209,46 @@ where
                             state_transition_name: Some(state_transition_name.to_string()),
                         })
                         .unwrap_or_else(error_to_internal_error_execution_result);
+
+                        #[cfg(test)]
+                        let execution_result =
+                            test_fault_injection::maybe_override(execution_result);
+
+                        if rollback_dropped_transitions {
+                            match &execution_result {
+                                StateTransitionExecutionResult::InternalError(_)
+                                | StateTransitionExecutionResult::UnpaidConsensusError(_) => {
+                                    // This transition will be stripped from the proposal
+                                    // (`TxAction::Removed`), so none of its writes may remain
+                                    // in the state the app hash is computed over. A rollback
+                                    // failure means the proposal can no longer match the
+                                    // block — fail it rather than continue on leaked state.
+                                    transaction.rollback_to_savepoint().map_err(|e| {
+                                        drive::grovedb::error::Error::StorageError(RocksDBError(e))
+                                    })?;
+                                }
+                                StateTransitionExecutionResult::SuccessfulExecution { .. }
+                                | StateTransitionExecutionResult::PaidConsensusError { .. } => {
+                                    // The transition stays in the block
+                                    // (`TxAction::Unmodified`), so its writes stay. Its
+                                    // savepoint is intentionally left on the stack (see
+                                    // `rollback_dropped_transitions` above: no
+                                    // pop-without-rollback exists, and at non-genesis heights
+                                    // the residue is inert).
+                                }
+                                StateTransitionExecutionResult::NotExecuted(_) => {
+                                    // Delayed to a later block (`TxAction::Delayed`) without
+                                    // having been executed: nothing was written since the
+                                    // savepoint, so rolling back and leaving it are
+                                    // equivalent. Leave it, like the kept outcomes above.
+                                    //
+                                    // Deliberately exhaustive: a new execution result variant
+                                    // must make an explicit savepoint decision here — the
+                                    // rollback classification must match the `TxAction`
+                                    // classification in `prepare_proposal`.
+                                }
+                            }
+                        }
 
                         // Store metrics
                         let elapsed_time = start_time.elapsed() + decoding_elapsed_time;
