+++ packages/rs-platform-version/src/version/drive_abci_versions/drive_abci_method_versions/mod.rs
@@ -1,7 +1,6 @@
 use versioned_feature_core::{FeatureVersion, OptionalFeatureVersion};
 
 pub mod v1;
-pub mod v10;
 pub mod v2;
 pub mod v3;
 pub mod v4;
